Validate service id is in the annotation name
